// packages/core/src/middleware/constants/MODULETYPE.d.ts
/**
 * @experimental  PASS ANIMATION
 */
export declare enum MODULETYPE {
    OBJECT3D = "object3D",
    CAMERA = "camera",
    LIGHT = "light",
    GEOMETRY = "geometry",
    TEXTURE = "texture",
    MATERIAL = "material",
    RENDERER = "renderer",
    SCENE = "scene",
    SPRITE = "sprite",
    CONTROLS = "controls",
    LINE = "line",
    MESH = "mesh",
    POINTS = "points",
    GROUP = "group",
    CSS3D = "css3D",
    CSS2D = "css2D",
    PASS = "pass",
    ANIMATION = "animation"
}
export declare const OBJECTMODULE: {
    camera: boolean;
    light: boolean;
    scene: boolean;
    sprite: boolean;
    line: boolean;
    mesh: boolean;
    points: boolean;
    group: boolean;
    css3D: boolean;
    css2D: boolean;
    object3D: boolean;
};